Refactor: 좋아요 토글 처리 시 redis 조회 및 저장 로직 제거
- postgreSQL 에서만 처리되게 변경
- netty worker thread count 별도 설정 코드 추가
- postgre-sql init.sql 중복으로 하나 제거
