+++ user-service/user-application/src/main/kotlin/com/jeffreyoh/userservice/application/config/UserServiceConfig.kt
@@ -1,17 +1,12 @@
 package com.jeffreyoh.userservice.application.config
 
 import com.jeffreyoh.userservice.application.port.`in`.CreatePostUseCase
+import com.jeffreyoh.userservice.application.port.`in`.SearchUseCase
 import com.jeffreyoh.userservice.application.port.`in`.TogglePostLikeUseCase
-import com.jeffreyoh.userservice.application.port.out.EventTrackerPort
-import com.jeffreyoh.userservice.application.port.out.RedisCommandPort
-import com.jeffreyoh.userservice.application.port.out.RedisReadPort
+import com.jeffreyoh.userservice.application.port.out.*
 import com.jeffreyoh.userservice.application.service.CreatePostService
 import com.jeffreyoh.userservice.application.service.SearchService
 import com.jeffreyoh.userservice.application.service.TogglePostLikeService
-import com.jeffreyoh.userservice.port.`in`.SearchUseCase
-import com.jeffreyoh.userservice.port.out.PostCommandPort
-import com.jeffreyoh.userservice.port.out.PostLikeCommandPort
-import com.jeffreyoh.userservice.port.out.PostSearchPort
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 
@@ -27,15 +22,11 @@ class UserServiceConfig {
     @Bean
     fun togglePostLikeUseCase(
         postLikeCommandPort: PostLikeCommandPort,
-        eventTrackerPort: EventTrackerPort,
-        redisReadPort: RedisReadPort,
-        redisCommandPort: RedisCommandPort
+        eventTrackerPort: EventTrackerPort
     ) : TogglePostLikeUseCase =
         TogglePostLikeService(
             postLikeCommandPort,
-            eventTrackerPort,
-            redisReadPort,
-            redisCommandPort
+            eventTrackerPort
         )
 
     @Bean
