+++ user-service/user-application/src/main/kotlin/com/jeffreyoh/userservice/application/service/CreatePostService.kt
@@ -2,7 +2,7 @@ package com.jeffreyoh.userservice.application.service
 
 import com.jeffreyoh.userservice.application.model.post.PostCommand
 import com.jeffreyoh.userservice.application.port.`in`.CreatePostUseCase
-import com.jeffreyoh.userservice.port.out.PostCommandPort
+import com.jeffreyoh.userservice.application.port.out.PostCommandPort
 import reactor.core.publisher.Mono
 
 class CreatePostService(
