+++ user-service/user-storage/src/test/kotlin/com/jeffreyoh/userservice/storage/adapter/outbound/redis/RedisReadAdapterTest.kt
@@ -12,7 +12,6 @@ import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.data.redis.core.ReactiveListOperations
 import org.springframework.data.redis.core.ReactiveStringRedisTemplate
 import reactor.core.publisher.Flux
-import reactor.core.publisher.Mono
 import reactor.test.StepVerifier
 
 @ExtendWith(MockKExtension::class)
@@ -33,38 +32,8 @@ class RedisReadAdapterTest {
 
     companion object {
         private const val RECENT_KEYWORD_KEY = "recent:search:user:%d"
-        private const val LIKE_CHECK_KEY = "like:user:%d:post:%d"
 
         fun getRecentKeywordKey(userId: Long) = RECENT_KEYWORD_KEY.format(userId)
-        fun getLikeCheckKey(userId: Long, postId: Long) = LIKE_CHECK_KEY.format(userId, postId)
-    }
-
-    @Test
-    fun `사용자의 좋아요 상태를 확인한다`() {
-        // given
-        val userId = 1L
-        val postId = 1L
-
-        val keySlot = slot<String>()
-        val expectedValue = "true"
-
-        every {
-            redisTemplate.opsForValue().get(capture(keySlot))
-        } returns Mono.just(expectedValue)
-
-        // when
-        val result = redisReadAdapter.getLikeCheck(userId, postId)
-
-        // then
-        StepVerifier.create(result)
-            .expectNext(true)
-            .verifyComplete()
-
-        verify(exactly = 1) { redisTemplate.opsForValue().get(keySlot.captured) }
-
-        val expectedKey = getLikeCheckKey(userId, postId)
-
-        assertThat(keySlot.captured).isEqualTo(expectedKey)
     }
 
     @Test
